Rename and add more functionality to maas tool
This change renames check_maas_plugins.py to rpc_maas_tool.py and adds
some functionality to mass delete checks and to print open alarms.
This is useful for a prod environments where you want to quickly see
what is currently down or for test environments where you need to
quickly delete all checks.

(cherry picked from commit dbffa3b6cd85285c4f915a9ec808ef1a3e7857e8)
(cherry picked from commit 9b0adafc85e80bcda940adbd5fbb12c1ec36f430)
